Don't do two ldconfigs on mips if we're not building kerberos. Many cross-tools cleanups: o remove mkdirs from the front of each of the rules, but instead have them   depend on new target cross-dirs, which makes all the directories o Add new target, cross-dirs, to create the directories and symlinks.  Before   the symlinks in /usr/cross/$TARGET/blah-unknown-openbsd2.4 were   never right. o KLUDGE for arc/pmax.  MACHINE_ARCH is mips for these machines, but they   differ from other machines in that they use mipsel-unknown-openbsd   for their gnu config string (failure to have the el causes big endian tools   to be built).
These fixes allow me to build a usable toolchain and libraries for
OpenBSD/arc on FreeBSD/i386 (with a few fixes to FreeBSD as well).  I
can also build a bootable kernel on the FreeBSD box.

I'd like to add a cross-build target shortly that will build
cross-tools, then will build and install stuff in DESTDIR with the
cross tools living in ${DESTDIR}/usr/cross/${TARGET}.