fix reconstruction performance. the old code used home-grown timers based upon hardcoded CPU speed values and an assumtion that the number of clock cycles was available. This is/was silly.
redone rf_GetNextReconEvent so that is now runs for 1/10th second
before sleeping for a short time (1/50th sec). Locally, this is using
about 25% of the CPU while rebuilding a disk in a four disk IDE RAID5
array. It was 22% of the way through when I last looked... much much
faster.

An even better way is sought - suggestions welcome. Lots of code that
the old routines relied on canm be harvested later.

Patches also being sent to Greg Oster @ NetBSD group.