Changes to pci and if_gm to support obtaining the ethernet hardware address from the pci bus. This is in preference to adding openfirmware code to the device drivers. If there was a known way of obtaining the ethernet hardware address from a eeprom or other methods that would be used, but the only known way to get this information for the Apple machines with if_gm is via openfirmware.
This modifies a previous mechanism that was used to obtain similar information
from different openfirmware systems, however the old mechanism would create
information such as media type. This information was hardcoded into
that code. Now the code only returns the actual address which is the
only informatin that openfirmware provides.