Apply recent changes in the i386 comment eradication code to all architectures using a list2sh.awk and related 'list' files (alpha, hp300, mac68k, mvme68k, mvme88k, mvmeppc, powerpc and vax).
list2sh.awk is modified so that in handling SPECIAL lines the tabs in
regular expressions are not lost to awk field parsing.

Any 'list' file trying to eliminate comments, using SPECIAL lines and
sed, now eliminates lines that

a) have '#' as the first non-blank, non-tab character, with at least
one blank or tab following the '#'.

b) have '#' as the first non-blank, non-tab character, with a newline
immediately following the '#'.

Previous b) was not being done, and only leading/following blanks were
being looked for due to the loss of tab characters in list2sh.awk.