Clean up the logic in install_url():
a) Eliminate extraneous cat invocations that are piped to
   grep by simply calling grep with the file name.
b) Eliminate the only use of isnumeric() by using a ksh'ism
   left exposed in sh: '+([0-9])'. Eliminate isnumeric().
c) Test for a http/ftp list index of '0' and treat as an out
   of range error rather than an IP address.
d) Reorganize to make smaller and more clear.

Add documentation to sh(1) for ksh patterns that are available.

ok millert@, miod@