Modify makedev so that it does not attempt to create network interface devices, to allow ask_which() to be used for interface configuration.
Rework network interface configuration logic to use ask_which() instead
of local code. This

1) May fix the problem noted by David Norman, where the prompts are
somehow broken, since no such problems have been noted with the
ask_which() code.

2) Shrinks the install output by a few lines.

3) Makes the install script smaller and clearer.

4) Makes interface configuration consistant with other 'device'
configurations.

5) Makes it impossible for a user to configure an interface more than
once.

Add the interface name to the 'IP address (or dhcp)' question. This
makes it more informative, especially if a '!' or '!foo' shell escape
has been used and overwritten the list of interfaces and the display of
the name of the one being configured.