The current hppa code would always ask for the root device, when running a kernel configured with "swap generic". This feature was relied upon by Theo de Raadt, so that he would have to go downstairs to type "sd0a\n\n" every time he would want to try a new kernel, effectively forcing his body to take exercise trim the fat.
No more! This commit adds code to correctly detect the root device, currently
either a network interface or an scsi block device. This is slightly tricky,
as all the information the PDC gives up is the controller device and its
numeric subpath, and we need to match the real device from that...
Tested and found to be working on ie and sd only so far.

This commit is brought to you by the so-called friends of Theo de Raadt.