Most of the non-US keyboard layouts define an ``AltGr'' key to access up to two extra symbols per key; usually, this key also serves as the compose key, if used with shift.
Because of this, accessing the fourth symbol of a key would work with a
sequence altgr-shift-key, but not with a sequence shift-altgr-key, since
it would trigger compose behaviour.

In order to be less confusing, let the AltGr behaviour win over the Compose
behaviour if AltGr would produce a different symbol. Since most of those
AltGr symbols are never part of compose rules, this is transparent.

Problem found by xsa@; ok deraadt@ otto@ xsa@