add support for multicast addresses. this adds the add/del multi handlers to the ioctl, and programs the multicast filter in tht_iff.
we are not using the perfect multicast filters, only the imperfect one. i
think there is a requirement that addresses in the perfect filter be in
order, but im not sure how to do that nicely (yet).

this has taken a long time because i dont know how to write c.